refactor: move reusable components to common folder
