+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 import './globals.css';
-import { Toaster } from '@/components/ui/toaster';
+import { Toaster } from '@/components/common/toaster';
 import AppLayout from '@/components/layout/AppLayout';
 import { Inter } from 'next/font/google';
 import { I18nProvider } from '@/i18n/I18nProvider';
