+++ src/components/ai/AutomatedTaskAssignmentFormCard.tsx
@@ -4,11 +4,11 @@ import { useState } from 'react';
 import { useForm, useFieldArray, type SubmitHandler } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import * as z from 'zod';
-import { Card, CardHeader, CardTitle, CardContent, CardFooter, CardDescription } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
+import { Card, CardHeader, CardTitle, CardContent, CardFooter, CardDescription } from '@/components/common/card';
+import { Button } from '@/components/common/button';
+import { Input } from '@/components/common/input';
+import { Label } from '@/components/common/label';
+import { Textarea } from '@/components/common/textarea';
 import { useToast } from '@/hooks/use-toast';
 import { assignTasksToEmployees, type AssignTasksToEmployeesInput, type AssignTasksToEmployeesOutput } from '@/ai/flows/assign-tasks-to-employees';
 import { UsersRound, PlusCircle, Trash2 } from 'lucide-react';
